refactor(questions): as system message
