--- src/components/chat/messages/question/index.scss
@@ -1,24 +1,0 @@
-@import 'src/styles/sizes';
-
-.question-wrapper {
-  background-color: var(--poll-wrapper-background-color);
-  border: $border-width-extra-small solid var(--poll-wrapper-color);
-  padding: $padding-small $padding-small 0 $padding-small;
-
-  [dir="ltr"] & {
-    margin: 0 $margin 0 0;
-  }
-
-  [dir="rtl"] & {
-    margin: 0 0 0 $margin;
-  }
-
-  .question-text,
-  .question-answer {
-    padding-bottom: $padding-small;
-  }
-
-  .question-label {
-    font-weight: var(--font-weight-semi-bold);
-  }
-}
